Major overhaul, including a public droid database

+++ bulk_id.php
@@ -16,12 +16,14 @@ if ($conn->connect_error) {
 
 function generateQR($id, $member_uid) {
     $path = "uploads/members/$member_uid/qr_code.jpg";
+    if (!file_exists($path)) {
     $link = "http://mot.astromech.info/id.php?id=".$id;
     $url = "https://chart.googleapis.com/chart?cht=qr&chld=L|1&chs=500x500&chl=".urlencode($link);
     echo "Generating QR Code: $url<br />";
     echo "Writing to path: $path<br />";
     $image = imagecreatefrompng($url);
     imagejpeg($image, $path, 75);
+    }
     return "Ok";
 }
 
@@ -60,7 +62,7 @@ if ($result->num_rows > 0) {
                 	mkdir("uploads/members/".$row['member_uid']);
         	}
 		if (!file_exists("uploads/members/".$row['member_uid']."/qr_code.jpg")) {
-			generateQR($row[badge_id], $row['member_uid']);
+			generateQR($row['badge_id'], $row['member_uid']);
 		}
